requirements.txt
asgiref==3.3.1
Django==3.1.4
django-optimistic-lock==1.0.0
django-sequences==2.5
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
